# fuse-service/src/main/resources/public/assets/swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0",
    "title": "Fuse API"
  },
  "basePath": "/fuse",
...
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json",
    "application/xml"
  ],
  "paths": {
...
    "/query/cypher": {
      "post": {
        "tags": [
          "query"
        ],
        "summary": "submit a cypher query request",
        "description": "submit a query to the engine",
        "parameters": [
          {
            "in": "body",
            "name": "<body>",
            "description": "Cypher Query json request",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CypherQueryRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns a Query summery with first page result data",
            "schema": {
              "$ref": "#/definitions/QueryInfo"
            }
          }
        }
      }
    },
    "/query/cypher/run": {
      "get": {
        "tags": [
          "query"
        ],
        "summary": "submit & run cypher query directly  ",
        "description": "submit & run a query to the engine",
        "parameters": [
          {
            "name": "cypher",
            "in": "query",
            "description": "cypher query",
            "required": true,
            "type": "string",
            "default": "match (e:Entity)-[hasEv:hasEvalue]->(v:Evalue), (e:Entity)-[hasRelEn:relatedEntity]->(e2:Entity) return *"
          },
          {
            "name": "ontology",
            "in": "query",
            "description": "ontology id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Returns a Query summery with first page result data",
            "schema": {
              "$ref": "#/definitions/QueryInfo"
            }
          }
        }
      }
    },
...
    "/query/{id}": {
      "get": {
        "tags": [
          "query"
        ],
        "summary": "Fetch query info by ID",
        "description": "Fetch query info by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "query ID.",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Returns <code>200</code> with a single query info or <code>404</code>",
            "schema": {
              "$ref": "#/definitions/QueryInfo"
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "tags": [
          "query"
        ],
        "summary": "Deletes a query by ID",
        "description": "Deletes a query by ID.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "query ID.",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "204": {
            "description": "A <code>204</code>"
          }
        }
      }
    },
    "/query/{id}/plan/json": {
      "get": {
        "tags": [
          "plan"
        ],
        "summary": "Fetch query plan by query ID",
        "description": "Fetch query plan by query ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "query ID.",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Returns <code>200</code> with a single query info or <code>404</code>",
            "schema": {
              "$ref": "#/definitions/QueryInfo"
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
  },
  "definitions": {
    "CypherQueryRequest": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int32"
        },
        "name": {
          "type": "string"
        },
        "query": {
          "type": "string"
        }
      },
      "example": {
        "id": "1",
        "name": "Q1",
        "ontology": "Knowledge",
        "query": "match (e:Entity)-[hasEv:hasEvalue]->(v:Evalue), (e:Entity)-[hasRelEn:relatedEntity]->(e2:Entity) return *"
      }
    },
...
    "QueryInfo": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int32"
        },
        "name": {
          "type": "string"
        }
      }
    },
...
  }
}